feat(validators): move report tool to validators index
Promote the validator report tool from /ethereum/validators/report to
the section index at /ethereum/validators, and remove the now-redundant
Validators subsection from the sidebar.

- Render IndexPage (the report tool) at the validators index route,
  carrying over the indices/from/to search param schema
- Delete the standalone report route and the card-based LandingPage that
  previously linked to it (and to a non-existent battle page)
- Update IndexPage useSearch/navigate references to target the index
  route instead of /report
- Drop the Validators nav array, its filter memo, and both mobile and
  desktop subsection blocks from the Sidebar
- Regenerate routeTree.gen.ts
