+++ src/pages/ethereum/validators/index.ts
@@ -1,2 +1,1 @@
 export { IndexPage } from './IndexPage';
-export { LandingPage } from './LandingPage';
