+++ src/routes/ethereum/validators/index.tsx
@@ -1,8 +1,16 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { LandingPage } from '@/pages/ethereum/validators';
+import { z } from 'zod';
+import { IndexPage } from '@/pages/ethereum/validators';
+
+const validatorsSearchSchema = z.object({
+  indices: z.string().optional(),
+  from: z.coerce.number().optional(),
+  to: z.coerce.number().optional(),
+});
 
 export const Route = createFileRoute('/ethereum/validators/')({
-  component: LandingPage,
+  component: IndexPage,
+  validateSearch: validatorsSearchSchema,
   beforeLoad: () => ({
     getBreadcrumb: () => ({ show: false }),
   }),
